requirements.txt
Django==3.1.3
mysqlclient==2.0.3
djangorestframework==3.12.2
django-filter==2.4.0
